Add tests for Search and App
