+++ package.json
@@ -32,6 +32,8 @@
     "enzyme-to-json": "^3.3.5"
   },
   "jest": {
-    "snapshotSerializers": ["enzyme-to-json/serializer"]
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ]
   }
 }
