--- src/components/favorites/Favorites.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-export class Favorites extends Component {
-  render() {
-    return (
-      <div>
-        
-      </div>
-    )
-  }
-}
-
-export default Favorites
