+++ src/tests/Header.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Header from './Header';
+import Header from '../layout/Header';
 import { shallow } from 'enzyme';
 
 describe('Header', () => {
